--- a/internal/utils/utils.go
+++ b/internal/utils/utils.go
@@ -1,6 +1,8 @@
 // +build !noasm
 
-// This file is here because Go cannot handle a package where all source
-// files are excluded due to build constraints, as is currently the case
-// for the arm64 build. Now there is always a valid package.
+// This file is here because the Go language specification requires that a
+// package has one or more source files. Now there is always a valid package.
+// Otherwise all source files would be excluded by the build constraints for
+// the arm64 build.
+// https://golang.org/ref/spec#Packages
 package utils