--- a/p751/arith_decl.go
+++ b/p751/arith_decl.go
@@ -1,4 +1,4 @@
-// +build amd64,!noasm
+// +build amd64,!noasm arm64,!noasm
 
 package p751
 
@@ -39,9 +39,7 @@ func fp751X2SubLazy(z, x, y *FpElementX2)
 //go:noescape
 func fp751Mul(z *FpElementX2, x, y *FpElement)
 
-// Function pointer that should point to one of the
-// fp751MontgomeryReduce implementations below.
-// When set, it performs Montgomery reduction: set z = x R^{-1} (mod 2*p).
+// Compute Montgomery reduction: set z = x R^{-1} (mod 2*p).
 // It may destroy the input value.
 //go:noescape
 func fp751MontgomeryReduce(z *FpElement, x *FpElementX2)