--- a/p751/arith_decl.go
+++ b/p751/arith_decl.go
@@ -39,7 +39,7 @@ func fp751X2SubLazy(z, x, y *FpElementX2)
 //go:noescape
 func fp751Mul(z *FpElementX2, x, y *FpElement)
 
-// Compute Montgomery reduction: set z = x R^{-1} (mod 2*p).
+// Compute Montgomery reduction: set z = x * R^{-1} (mod 2*p).
 // It may destroy the input value.
 //go:noescape
 func fp751MontgomeryReduce(z *FpElement, x *FpElementX2)