@henrydcase 2018-09-13
This PR needs to be based on PR#14 in order to make it possible to test against reference implementation.