@Ko- 2018-09-17
Rebased on #14. I resolved the thing where `fp751MontgomeryReduce` has multiple implementations on `amd64` but only one on `arm64` by splitting the declarations to separate files with build tags and only have the function pointer left in `arith_decl.go`.