@Ko- 2018-09-23
I didn't know about those tools, thanks for the suggestion!